refactor: extract TestOptionsMonitor to shared test helper
